requirements.txt
streamlit
openai
langchain
PyMuPDF
python-dotenv

